+++ crates/ethcore/types/Cargo.toml
@@ -6,7 +6,6 @@ authors = ["Parity Technologies <admin@parity.io>"]
 
 [dependencies]
 ethereum-types = "0.4"
-ethjson = { path = "../../ethjson" }
 ethkey = { path = "../../accounts/ethkey" }
 heapsize = "0.4"
 keccak-hash = "0.1"
