+++ crates/ethjson/src/blockchain/transaction.rs
@@ -41,7 +41,7 @@ pub struct Transaction {
 
 pub type AccessList = Vec<AccessListItem>;
 
-#[derive(Debug, PartialEq, Deserialize)]
+#[derive(Debug, PartialEq, Deserialize, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct AccessListItem {
     pub address: H160,
