+++ crates/ethjson/src/lib.rs
@@ -14,7 +14,9 @@
 // You should have received a copy of the GNU General Public License
 // along with OpenEthereum.  If not, see <http://www.gnu.org/licenses/>.
 
+extern crate common_types as types;
 extern crate ethereum_types;
+extern crate ethkey;
 extern crate rustc_hex;
 extern crate serde;
 extern crate serde_json;
